Методы разнесены по классам и функциональности, убрана конкатенация (заменил на printf) Исправлены возвращающие типы значений Запись  и считывание с файла сделал универсальным без выбора типа
